Add files via upload
Construction of haploid Kauffman NK model and simulation of random walks within the landscape. Here the whole landscape is generated.

For NK model, please read the paper: Kauffman, S.A. and Weinberger, E.D., 1989. The NK model of rugged fitness landscapes and its application to maturation of the immune response. Journal of theoretical biology, 141(2), pp.211-245.